requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
cachetools==5.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
comm==0.1.2
cryptography==41.0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.2
dnslib==0.9.23
dnspython==2.4.2
ecdsa==0.18.0
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
gcalcli==4.3.0
gdata==2.0.18
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
localstack==2.2.0
localstack-core==2.2.0
localstack-ext==2.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
meson==1.1.1
mistune==2.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.3
notebook_shim==0.2.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
plux==1.4.0
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.24.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rsa==4.9
semver==3.0.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SpeechRecognition==3.10.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
stack-data==0.6.2
stevedore==5.1.0
tabulate==0.9.0
tailer==0.4.1
terminado==0.17.1
tinycss2==1.2.1
tk==0.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
y-py==0.5.9
ypy-websocket==0.8.2
